requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==43.0.1
python-telegram-bot==21.6
MetaTrader5==5.0.45
tenacity==9.0.0
orjson==3.10.7
